mobile: AccountScreen: copy link to account
