+++ apps/daimo-mobile/src/view/shared/useSwipeUpDown.tsx
@@ -73,7 +73,7 @@ export function useSwipeUpDown({
         ref={bottomSheetRef}
         itemMini={itemMini}
         itemFull={itemFull}
-        swipeHeight={(screenDimensions.height / Math.PI) | 0}
+        swipeHeight={(screenDimensions.height / 3.5) | 0}
         onShowFull={onOpenTransactionsModal}
         onShowMini={onCloseTransactionsModal}
         disabled={disabled}
